chore(surveys): address review feedback on survey bar charts
- drop "Quill" from the chart file/component names
- remove the data-derived useMemos (needless for this little data)
- extract the chart tooltips into their own components
- drop the pointless click-handler wrapper functions
