--- frontend/src/scenes/surveys/components/question-visualizations/MultipleChoiceBarChartQuill.tsx
@@ -1,146 +1,0 @@
-import { useValues } from 'kea'
-import { useMemo } from 'react'
-
-import { BarChart, ValueLabels } from '@posthog/quill-charts'
-import type { BarChartConfig, PointClickData, Series, TooltipContext } from '@posthog/quill-charts'
-
-import { buildTheme } from 'lib/charts/utils/theme'
-import { computeBarColors } from 'scenes/surveys/components/question-visualizations/questionVizTransforms'
-import { CHART_INSIGHTS_COLORS } from 'scenes/surveys/components/question-visualizations/util'
-
-import { themeLogic } from '~/layout/navigation-3000/themeLogic'
-import { ChoiceQuestionResponseData } from '~/types'
-
-const CATEGORY_LABEL_WIDTH = 280
-
-interface TooltipContextData {
-    rank: number
-    respondentPercentage: string
-    selectionPercentage: string
-}
-
-interface Props {
-    chartData: ChoiceQuestionResponseData[]
-    totalResponses: number
-    activeChoiceLabel: string | null
-    highlightedChoiceLabel: string | null
-    tooltipContextByIndex: TooltipContextData[]
-    onBarClick: (index: number) => void
-}
-
-export function MultipleChoiceBarChartQuill({
-    chartData,
-    totalResponses,
-    activeChoiceLabel,
-    highlightedChoiceLabel,
-    tooltipContextByIndex,
-    onBarClick,
-}: Props): JSX.Element {
-    const { isDarkModeOn } = useValues(themeLogic)
-    // isDarkModeOn invalidates the memo so buildTheme() re-reads CSS vars on dark-mode toggle.
-    const theme = useMemo(() => buildTheme(), [isDarkModeOn])
-
-    const barColors = useMemo(() => {
-        const baseColors = chartData.map((_, i) => CHART_INSIGHTS_COLORS[i % CHART_INSIGHTS_COLORS.length])
-        return computeBarColors(
-            baseColors,
-            chartData.map((d) => d.label),
-            highlightedChoiceLabel,
-            !!activeChoiceLabel
-        )
-    }, [activeChoiceLabel, chartData, highlightedChoiceLabel])
-
-    const series = useMemo<Series[]>(
-        () => [
-            {
-                key: 'multiple-choice',
-                label: 'Number of responses',
-                data: chartData.map((d) => d.value),
-                color: barColors[0],
-                bars: chartData.map((_, i) => ({ color: barColors[i] })),
-            },
-        ],
-        [chartData, barColors]
-    )
-
-    // Synthetic band keys keep bars distinct even if two choices share a label; the choice text is
-    // rendered through the category-axis formatter instead.
-    const labels = useMemo(() => chartData.map((_, i) => String(i)), [chartData])
-
-    const config = useMemo<BarChartConfig>(
-        () => ({
-            hideXAxis: true,
-            axisOrientation: 'horizontal',
-            maxCategoryLabelWidth: CATEGORY_LABEL_WIDTH,
-            xTickFormatter: (_label, index) => chartData[index]?.label ?? '',
-            bars: { minBandSize: 32, bandPadding: 0.4 },
-        }),
-        [chartData]
-    )
-
-    const valueLabelFormatter = (value: number): string => {
-        const total = totalResponses || 1
-        const percentage = ((value / total) * 100).toFixed(1)
-        return `${value} (${percentage}%)`
-    }
-
-    const renderTooltip = (ctx: TooltipContext): JSX.Element | null => {
-        const optionLabel = chartData[ctx.dataIndex]?.label ?? ctx.seriesData[0]?.series.label ?? ''
-        const tooltipContext = tooltipContextByIndex[ctx.dataIndex]
-        const value = ctx.seriesData[0]?.value ?? 0
-
-        let inspectLabel = 'Double click to filter'
-        if (activeChoiceLabel && optionLabel === activeChoiceLabel) {
-            inspectLabel = 'Click to clear filter'
-        } else if (activeChoiceLabel) {
-            inspectLabel = 'Click to switch filter'
-        }
-
-        if (!tooltipContext) {
-            return (
-                <div className="bg-surface-primary border rounded-md shadow-md px-3 py-2 text-sm">
-                    <span className="font-medium">{optionLabel}</span>
-                </div>
-            )
-        }
-
-        return (
-            <div className="bg-surface-primary border rounded-md shadow-md px-3 py-2 text-sm">
-                <div className="flex items-center gap-2 leading-tight">
-                    <span className="font-semibold">{optionLabel}</span>
-                    <span className="text-xs text-muted-alt">
-                        #{tooltipContext.rank} of {chartData.length}
-                    </span>
-                </div>
-                <div className="text-xs text-secondary leading-tight mt-0.5">
-                    <span className="font-semibold tabular-nums text-primary">{value}</span> responses
-                    <span className="mx-1 text-muted-alt">•</span>
-                    <span className="font-semibold text-primary">{tooltipContext.respondentPercentage}%</span>{' '}
-                    respondents
-                    <span className="mx-1 text-muted-alt">•</span>
-                    <span className="font-medium text-primary">{tooltipContext.selectionPercentage}%</span> of all
-                    selected options
-                </div>
-                <div className="text-xs text-muted mt-1">{inspectLabel}</div>
-            </div>
-        )
-    }
-
-    const handlePointClick = ({ dataIndex }: PointClickData): void => onBarClick(dataIndex)
-
-    return (
-        <div className="pl-4">
-            <BarChart
-                series={series}
-                labels={labels}
-                config={config}
-                theme={theme}
-                tooltip={renderTooltip}
-                onPointClick={handlePointClick}
-                dataAttr="survey-multiple-choice"
-            >
-                <ValueLabels valueFormatter={valueLabelFormatter} />
-            </BarChart>
-        </div>
-    )
-}
